+++ packages/wonder-blocks-testing/CHANGELOG.md
@@ -1,5 +1,34 @@
 # @khanacademy/wonder-blocks-testing
 
+## 4.0.4
+
+### Patch Changes
+
+-   Updated dependencies [580141ed]
+    -   @khanacademy/wonder-blocks-data@8.0.2
+
+## 4.0.3
+
+### Patch Changes
+
+-   Updated dependencies [e5fa4d9e]
+    -   @khanacademy/wonder-blocks-data@8.0.1
+
+## 4.0.2
+
+### Patch Changes
+
+-   Updated dependencies [1385f468]
+-   Updated dependencies [0720470e]
+-   Updated dependencies [0720470e]
+-   Updated dependencies [cf9ed87f]
+-   Updated dependencies [b882b082]
+-   Updated dependencies [0720470e]
+-   Updated dependencies [75c10036]
+-   Updated dependencies [a85f2f3a]
+-   Updated dependencies [0720470e]
+    -   @khanacademy/wonder-blocks-data@8.0.0
+
 ## 4.0.1
 
 ### Patch Changes